Enable unused_imports, unused_qualifications and trivial_numeric_casts
The first was allowed at the crate root and in four more files; the other
two are allow-by-default. All three are residue of moving C++ files across
rather than anything the Rust needs: 434 imports nothing used, 411 paths
spelled out where the name was already in scope, and 249 casts from a type
to itself.

Most of it is `cargo fix`. Three classes needed hands:

  * `cargo fix --lib` compiles without cfg(test), so it stripped imports
    that only the in-module unit tests reach. Those move into their
    `mod tests` rather than back to the top of the file, where the
    non-test build would call them unused in turn. Same shape in the
    cfg-gated SIMD sets, which the default aarch64 build never compiles:
    x86_64 and wide are now checked per feature and per target.

  * `trivial_numeric_casts` carries no machine-applicable suggestion, so
    the 249 are stripped from the lint's own spans. Seven were wrapped in
    parentheses that existed only to hold the cast and had to go with it,
    or `unused_parens` (denied) rejects the result.

  * The lint is reported per macro *expansion*, and `check_ref_idx!` in
    both B-slice parsers is expanded with `$iref` as `i8` at one site and
    `i32` at another: dropping the cast there stops compiling. The
    `$listIdx` casts really were no-ops and go; the `$iref` one becomes
    `i32::from`, which is correct for both without a cast at all. An
    `#[allow]` on the macro would not have worked anyway — the level is
    resolved at the expansion site.

tests/common/mod.rs keeps two scoped allows on its convenience re-exports,
for the reason `dead_code` is allowed there: each consumer is its own
crate and pulls in one of them, or neither.

Clean on default / wide / scalar / wide+scalar, debug and release,
aarch64 and x86_64. Byte-identical: diffharness 847/847 over
st qp ps ltr sl dl mt def.

+++ rust/crates/openh264-rs/src/decoder/mv_pred.rs
@@ -44,7 +44,6 @@
 )]
 #![forbid(unsafe_code)]
 
-use std::ffi::c_void;
 use crate::safe::mb_grid::MbArray;
 
 // ============================================================================
@@ -180,7 +179,7 @@ pub use crate::decoder::decoder_core::{SSlice, SLayerInfo, DqLayerState};
 
 
 pub use crate::decoder::decoder_context::{SRefPic};
-use crate::decoder::decoder_context::{SliceCtx, active_pps, active_sps, pps_of, sps_of};
+use crate::decoder::decoder_context::SliceCtx;
 pub use crate::decoder::decoder_context::{
     SWelsDecoderContext, PicRefs, ref_id,
 };
@@ -831,7 +830,7 @@ pub fn PredMvBDirectSpatial(
         }
         UpdateP16x16DirectCabac(pCurDqLayer);
         for listIdx in 0..2 {
-            UpdateP16x16MotionInfo(pCurDqLayer, Some(&mut *pDec), listIdx, ref_idx[listIdx as usize], &iMvp[listIdx as usize]);
+            UpdateP16x16MotionInfo(pCurDqLayer, Some(&mut *pDec), listIdx, ref_idx[listIdx], &iMvp[listIdx]);
             UpdateP16x16MvdCabac(pCurDqLayer, &pMvd, listIdx as i32);
         }
     } else {
